+++ src/components/Channel/ThreadCard.js
