+++ src/components/Thread/InputModal.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import ReactModal from 'react-modal';
-import UploadPicButton from './UploadPicButton';
+import UploadPicButton from '../UploadPicButton';
 import { makeStyles } from '@material-ui/core/styles';
 import IconButton from '@material-ui/core/IconButton';
 import Button from '@material-ui/core/Button';
