+++ src/containers/Login.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import { makeStyles, useTheme } from '@material-ui/core/styles';
-import { TextFields, Buttons } from '../components/LoginForm';
+import { TextFields, Buttons } from '../components/Login/LoginForm';
 import firebase from 'firebase/app';
 import 'firebase/auth';
